mmc: core: Add CAP2_FULL_PWR_CYCLE cap check
Even if, at the moment, we don't support Kitakami (yet?), this
is a critical fix for its hw problems: without this, the eMMC
will get permanently damaged.

As for Tone, this should be only for a performance improvement.
The eMMC damage, like Kitakami, has not been verified on this
platform.

Change-Id: Ib9402580a8f33472ae91939134ef104a61be5b76